models: Group's are now related to any kind of Contactable
The Group model is related to Role which used to be in turn related
to Person.  Now the Role is related to Contactable, which enables
Organisation, Member or Person to be in a Group.
The `current_groups' method is now moved from Person to Contactable
as it applies to all of them.  This method returns a string with
the list of groups a Contactable belongs to.
A new property Contactable.type_str returns a string with the type
of Contactable for the current instance.

SQL: fix-013.sql
(rename cams_role.person_id -> cams_role.contactable_id)