models: sync Member entries status when person/org status changes
When the status of either the Person or Organisation fields of a
Member entry has a "DISABLED" status, or both, then the Member
entry must also be disabled.  If one of them is "NEW" then the
Member must also be "NEW".  Finally, When both are "ACTIVE", then
the Member must be active.